Andrew Im (Changes Made)
Created 2 Scripts:
PlayerHealth (Contained in Player scripes) & HealthBar (Contained in scripes)

Added Music: Unity (Which is in Hierarchy as a Audio Source. Plays when it is started)

Basic Summary:
I added a Player health bar attached to the camera. It can't be seen behind the cube though. By pressing on the space bar the health of the player decreases (is in the script of PlayerHealth).